refactor(createDataGrid): simplify adapters, layout, and factory
- Extract shared applyOrder helper, eliminating duplicate ordering
  logic between ClientGridAdapter and VirtualGridAdapter
- Fix O(n*m) order.includes() → Set-based O(n) lookup in adapters
  and ordering.apply()
- Simplify layout.ts columns computed to reuse pinned regions
  instead of double-computing resolved()
- Fix editable column filter to match editing.ts guard
- Remove unused RowSpanningOptions import and numbered comments
- Clean up test mock imports
